+++ changelog.md
@@ -1,3 +1,4 @@
 ## Version 3.2
 - Fixed bug with entities being misplaced after map size is expanded.
 - Made currently open file show up in the window title.
+- Status bar will show cursor grid coordinates when nothing else is showing.